fix: 修复对象池统计信息中"using_obj_rate(对象使用率)"统计为0的bug

+++ llbc/include/llbc/core/objpool/ObjPoolInl.h
@@ -442,7 +442,7 @@ LLBC_Json::Value LLBC_TypedObjPool<Obj>::GetStatistics(LLBC_Json::MemoryPoolAllo
     stat.AddMember("using_obj_count", static_cast<uint32>(_usingObjCount), jsonAlloc);
     // - using_obj_rate.
     stat.AddMember("using_obj_rate",
-                   objCount != 0 ? static_cast<uint32>(_usingObjCount) / objCount : 0.0,
+                   objCount != 0 ? static_cast<double>(_usingObjCount) / objCount : 0.0,
                    jsonAlloc);
     // - reusable_obj_count.
     stat.AddMember("reusable_obj_count", _reusableObjCount, jsonAlloc);
